Compartmentalized functions
optimization.py was getting very big...
