+++ tests/test_opt/model_losses.csv
@@ -1,0 +1,13 @@
+timestamp,iteration,loss,train_r2,train_mae,val_r2,val_mae
+2025-07-10T14:33:17.266909,1,1.2782506942749023,0.9988331838785223,0.0064253697352580266,-0.18656487792187337,0.2692629081324415
+2025-07-10T14:33:17.398386,2,1.4863361120224,0.9980250104710199,0.008067032423475198,-0.7203839447344194,0.32789269638875834
+2025-07-10T14:33:17.551845,3,1.3569815158843994,0.9937639576760366,0.01639250684466656,-0.505502848203542,0.29788181632561916
+2025-07-10T14:33:17.708919,4,1.444853961467743,0.9969762428621226,0.009672699413422803,-0.2919633570052911,0.28781782961429697
+2025-07-10T14:33:17.860454,5,1.6108241081237793,0.9925770815971215,0.00998771738507931,-0.3888105248768845,0.3103277640062446
+2025-07-10T14:33:18.022945,6,1.3624882698059082,0.9825303718771008,0.018834104031324714,-0.5064607870339402,0.27793360915259346
+2025-07-10T14:33:18.215722,7,1.3605414628982544,0.9906205299404162,0.010418650246171871,-0.2172746656145499,0.27464786321346285
+2025-07-10T14:33:18.376092,8,1.1818535327911377,0.9704767133044099,0.02669772415066167,-0.6081292823027482,0.28730273307350995
+2025-07-10T14:33:18.545833,9,1.3180419206619263,0.9707858233072904,0.025934968454170922,-0.4178137442933334,0.29187198540669224
+2025-07-10T14:33:18.727060,10,1.5783318281173706,0.9802321412548878,0.020638880357123517,-0.9179255550535133,0.3295301359112014
+2025-07-10T14:33:18.899470,11,1.2442132234573364,0.9647655071821605,0.02682115984596532,-0.358398328758331,0.27067014668835204
+2025-07-10T14:33:19.077698,12,1.4265530109405518,0.9582444780035249,0.029661490199902864,-0.5367114850218253,0.3033226254025353
